+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.5.0 (unreleased)
 
 - Added experimental support for OpenSearch
+- Added support for synonyms in Japanese
 
 ## 4.4.4 (2021-03-12)